chore(backend): rename football_identity.rs to identity_upgrade.rs
Rename the ofm_core module from football_identity to identity_upgrade
to better reflect its purpose as an identity field upgrade mechanism.

Updates module declaration in lib.rs and all crate references.

Closes #59
